+++ modules/tools/agic/variables.tf
@@ -308,6 +308,5 @@ variable "enable_agic" {
 variable "agic_helm_version" {
   description = "Version of Helm chart to deploy"
   type        = string
-  # TODO : Watch for the release of 1.2.0 final version
-  default = "1.2.0-rc2"
+  default     = "1.2.0"
 }