+++ modules/tools/velero/locals-velero.tf
@@ -23,15 +23,15 @@ EOF
     "configuration.backupStorageLocation.bucket"                = try(azurerm_storage_container.velero.0.name, "")
     "configuration.backupStorageLocation.config.resourceGroup"  = try(azurerm_storage_account.velero.0.resource_group_name, "")
     "configuration.backupStorageLocation.config.storageAccount" = try(azurerm_storage_account.velero.0.name, "")
-    "configuration.backupStorageLocation.name"                  = "azure"
+    "configuration.backupStorageLocation.name"                  = "default"
     "configuration.provider"                                    = "azure"
     "configuration.volumeSnapshotLocation.config.resourceGroup" = try(var.aks_nodes_resource_group_name, "")
-    "configuration.volumeSnapshotLocation.name"                 = "azure"
+    "configuration.volumeSnapshotLocation.name"                 = "default"
     "credentials.existingSecret"                                = try(kubernetes_secret.velero.0.metadata.0.name, "")
     "credentials.useSecret"                                     = "true"
     "deployRestic"                                              = "false"
     "env.AZURE_CREDENTIALS_FILE"                                = "/credentials"
-    "metrics.enabled"                                           = "false"
+    "metrics.enabled"                                           = "true"
     "rbac.create"                                               = "true"
     "schedules.daily.schedule"                                  = "0 23 * * *"
     "schedules.daily.template.includedNamespaces"               = "{'*'}"
@@ -40,13 +40,14 @@ EOF
     "serviceAccount.server.create"                              = "true"
     "snapshotsEnabled"                                          = "true"
     "initContainers[0].name"                                    = "velero-plugin-for-azure"
-    "initContainers[0].image"                                   = "velero/velero-plugin-for-microsoft-azure:master"
+    "initContainers[0].image"                                   = "velero/velero-plugin-for-microsoft-azure:main"
     "initContainers[0].volumeMounts[0].mountPath"               = "/target"
     "initContainers[0].volumeMounts[0].name"                    = "plugins"
     "image.repository"                                          = "velero/velero"
-    "image.tag"                                                 = "master"
+    "image.tag"                                                 = "v1.4.0"
     "image.pullPolicy"                                          = "IfNotPresent"
     "podAnnotations.aadpodidbinding"                            = local.velero_identity_name
+    "podLabels.aadpodidbinding"                                 = local.velero_identity_name
   }